refactor(InvalidPathException.java) : 경로 조회 시 예외처리 로직 변경
