+++ src/main/java/wooteco/subway/exception/badrequest/InvalidPathException.java
@@ -1,14 +1,17 @@
-package wooteco.subway.path.application;
+package wooteco.subway.exception.badrequest;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class InvalidPathException extends RuntimeException {
-    public InvalidPathException() {
-    }
+public class InvalidPathException extends BadRequest {
+
 
     public InvalidPathException(String message) {
         super(message);
+    }
+
+    public InvalidPathException(String message, Throwable cause) {
+        super(message, cause);
     }
 }
